Filter

+++ variant/vcf.py
@@ -271,4 +271,4 @@ def get_vcf_population_allelic_frequencies(caf):
     :return: list; (n_alleles); of .VCF population allelic frequencies
     """
 
-    return [int(a_caf) for a_caf in caf.split(',')]
+    return [float(a_caf) for a_caf in caf.split(',')]
